src/requirements.txt
boto3
sentence-transformers
accelerate
stomp.py
